Load vnode.nodeName only once.
Since idiff is invoked on all kinds of virtual nodes, the property
access vnode.nodeName is usually megamorphic in any realistic workload,
so it's better for the JS engine to load the property only once.

Also apply the String constructor only once to the vnode.nodeName,
because the String constructor can have arbitrary side-effects (since it
can call out to Symbol.toPrimitive, valueOf or toString), and thus
requires magic on the engines' side to optimize this away, which doesn't
always kick in (and is costly).